lib: use pathname for record file
Previously the `RecordBuilder` used in `host_log()` to construct a `Log`
hardcoded the record's `file()` as `Some("app.rs")`. It seems better in
practice to use `&pathname`, the value that is provided to
`module_path()`.

Before this change when using `tracing-log` to adapt the `Log` record to
the `tracing` ecosystem an event like the following was produced:

```
2024-07-23T15:31:24.376313Z DEBUG init: app.rs:10: User Python Loaded!
```

Afterwards, it produces an event like:

```
2024-07-23T15:50:38.273451Z DEBUG init: /home/daniel/.config/pup/user_test.py:10: User Python Loaded!
```